+++ entroppy/__main__.py
@@ -5,7 +5,8 @@ from loguru import logger
 from entroppy.cli import create_parser
 from entroppy.core import load_config
 from entroppy.processing import run_pipeline
-from entroppy.utils import DebugTypoMatcher, setup_logger
+from entroppy.utils.debug import DebugTypoMatcher
+from entroppy.utils.logging import setup_logger
 
 
 def main():